pyproject.toml
[project]
name = "lhammai-cli"
version = "0.1.0"
description = "Interact with any LLM from your terminal"
readme = "README.md"
authors = [
    { name = "Dimitris Poulopoulos", email = "dimitris.a.poulopoulos@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "dotenv>=0.9.9",
    "halo>=0.0.31",
    "pydantic-settings>=2.10.1",
    "rich>=14.1.0",
]

[project.scripts]
lhammai = "lhammai_cli:main"

[build-system]
requires = ["uv_build>=0.8.11,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
ollama = [
    "any-llm-sdk[ollama]>=0.13.1",
]
